fix: Unit tests for HOC refactor
