+++ src/components/ui/List/Item.js
@@ -6,7 +6,7 @@ import React from 'react';
 import { toUnits } from 'styles';
 import { useTheme } from 'themes';
 
-export const componentClassName = 'Nautilus-ListItem';
+export const ComponentClassName = 'Nautilus-ListItem';
 
 export const Item = (props) => {
   const { children, className, ...otherProps } = props;
@@ -15,7 +15,7 @@ export const Item = (props) => {
 
   return (
     <li
-      className={classnames(componentClassName, className)}
+      className={classnames(ComponentClassName, className)}
       css={css`
         position: relative;
         margin: 0 0 ${toUnits(theme.spacing.padding.xSmall)};
