+++ src/components/ui/Tabs/index.test.js
@@ -1,4 +1,4 @@
-import React, { createRef } from 'react';
+import React from 'react';
 
 import { axe, fireEvent, render, wait, waitForElement } from 'utils/testing';
